Move experimental menu commands into Developer auth module
